Move AppleSignInControllerAuthAdapterTests to its own file
